+++ src/main/kotlin/io/github/teambluemods/healthmod/items/SyringeItem.kt
@@ -1,0 +1,45 @@
+/*
+ * Copyright (c) 2020, 2021 Team Blue.
+ *
+ * This file is part of HealthMod Fabric.
+ *
+ * HealthMod Fabric is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * HealthMod Fabric is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with HealthMod Fabric.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package io.github.teambluemods.healthmod.items
+
+import io.github.blueminecraftteam.healthmod.registries.ItemRegistries
+import io.github.teambluemods.healthmod.util.extensions.isServer
+import net.minecraft.entity.damage.DamageSource
+import net.minecraft.entity.player.PlayerEntity
+import net.minecraft.item.Item
+import net.minecraft.item.ItemStack
+import net.minecraft.util.Hand
+import net.minecraft.util.TypedActionResult
+import net.minecraft.world.World
+
+class SyringeItem(settings: Item.Settings): Item(settings) {
+    override fun use(world: World, user: PlayerEntity, hand: Hand): TypedActionResult<ItemStack> {
+        if (world.isServer) {
+            val stackInHand = user.getStackInHand(hand)
+
+            if (world.isServer) {
+                val setStackInHand = user.giveItemStack(ItemRegistries.SYRINGE.defaultStack)
+                user.damage(DamageSource.GENERIC, 0.5f)
+            }
+        }
+
+        return super.use(world, user, hand)
+    }
+}
